% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consulta_modelos.R
\name{consulta_modelos}
\alias{consulta_modelos}
\title{consulta os modelos disponiveis por marca da tabela fipe}
\usage{
consulta_modelos(marca)
}
\arguments{
\item{marca}{código da marca importada pela fun pega_marcas}
}
\description{
consulta os modelos disponiveis por marca da tabela fipe
}
