% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pega_ano_modelo.R
\name{pega_ano_modelo}
\alias{pega_ano_modelo}
\title{pega ano dos modelos dos carros da tabela fipe}
\usage{
pega_ano_modelo(modelo, codigo_marca)
}
\arguments{
\item{modelo}{modelo do carro}

\item{codigo_marca}{codigo_marca}
}
\description{
pega ano dos modelos dos carros da tabela fipe
}
