% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pega_marcas.R
\name{pega_marcas}
\alias{pega_marcas}
\title{pega_marcas disponveis por titulo de veiculo na tabela fipe}
\usage{
pega_marcas(tipo_veiculo = 1)
}
\arguments{
\item{tipo_veiculo}{identificar o tipo de veiculo se 1, 2 ou 3}
}
\description{
pega_marcas disponveis por titulo de veiculo na tabela fipe
}
