% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pega_tabela_fipe.R
\name{pega_tabela_fipe}
\alias{pega_tabela_fipe}
\title{baixa os dados da tabela fipe}
\usage{
pega_tabela_fipe(dataset_com_modelos)
}
\arguments{
\item{dataset_com_modelos}{dataset importando os modelos usar a fun consulta_modelos()}
}
\description{
baixa os dados da tabela fipe
}
